feat: adjust package name to renamed directory
