+++ cli/internal/graph/renderer/render.go
@@ -1,4 +1,4 @@
-package displaymanager
+package renderer
 
 import (
 	"bytes"
