+++ cli/internal/graph/renderer/render_test.go
@@ -1,4 +1,4 @@
-package displaymanager
+package renderer
 
 import (
 	"bytes"
